backend/requirements.txt
asgiref==3.5.2
astroid==2.12.12
colorama==0.4.6
dill==0.3.6
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.3
PyJWT==2.6.0
pylint==2.15.5
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.6
wrapt==1.14.1
